+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-07-27)
+# 0.0.0 (2021-07-28)
 
 ### Bug Fixes
 
@@ -70,6 +70,7 @@
 - **algorithm/math:** add subarraySum ([69730c2](https://github.com/air-supply94/js-algorithm/commit/69730c229b2b325e902a04ce93692e72bd0d14e4))
 - **algorithm/math:** add trailingZeroes ([cefab4c](https://github.com/air-supply94/js-algorithm/commit/cefab4c03594616e8c8ff8ed447a4bdd1febff3d))
 - **algorithm/math:** add WSRARes ([d6589c4](https://github.com/air-supply94/js-algorithm/commit/d6589c43f9c6dd469b0a642b4df9f43a136de3b8))
+- **algorithm/math/bits:** add exchangeBits ([ee1afff](https://github.com/air-supply94/js-algorithm/commit/ee1afffac76bf5076a255baf5ff4569e088d8ab6))
 - **algorithm/other:** add isRectangleCover ([dae0405](https://github.com/air-supply94/js-algorithm/commit/dae04055889731282e7e870c53f4594a28dda527))
 - **algorithm/other:** add isSubsequence ([f111959](https://github.com/air-supply94/js-algorithm/commit/f111959d6a19722ba218580f70e5ac2ab82e705b))
 - **algorithm/other:** add oneEditAway ([8462de8](https://github.com/air-supply94/js-algorithm/commit/8462de8cc4659d51163106296a16368ba152c122))